+++ NEWS.md
@@ -1,5 +1,13 @@
 # 📰 News
 
+## v0.10.0
+
+- WARNING: Breaking API, CLI & Config changes.
+- Write metadata from the command line.
+- Optional PDF support. Install as `comicbox[pdf]`
+- Python version now >=3.9
+- Extract a range of pages from the cli.
+
 ## v0.9.1
 
 - Removed dependence on Python 3.11